Check that there are results before computing 'sinceId'
If there are no existing results, then the `Math.max()` call returns
`-Infinity`, which is not a valid id for the API.

To replicate the issue: Search for a phrase that won't exist in the database, then observe that while the initial search returns a 200 and an empty list of results, subsequent auto-update `fetch` requests receive a 400.

To test the fix: Make the same search and observer that auto-update
requests receive a 200 with empty results.