Request stories from the last n*1.2 seconds since last poll
In conversation with Reuters they said that there's currently no way of
requesting just the stories that have changed since last request
('endCursor' shouldn't be relied upon for this).

So we should search for a time period that overlaps with our last
request, and page through the results. The ingestion lambda handles
deduplication now, so no need to deduplicate here unless we spot
performance issues, for instance.