Remove ability to search by subject
The aim is for all useful subject data to be captured in the
'categoryCodes' field, which is more standardised and easier to search
given that it's not wrapped in a JSON column.

Unless there are serious issues we should aim to fix forward any issues
with categoryCode search that come up.

I'm leaving subject info in the API response at it could be helpful for
debugging, but in theory we shouldn't have any need for this field after
ingestion.