ingestion: fix bad tab literal characters
We receive a bunch of input documents where the body text field contains a literal tab character
(rather than the usual \t escaped version), which is invalid in JSON. Make a best effort to fix it up,
but this is a little tricky as the tab characters are valid outside of JSON strings (ie. as indentation
around the object literal).