Ingestion goes via intermediate S3 storage step
Both the fingerpost-queueing-lambda and the poller-lambdas write content
to S3 and then queue work for the ingestion-lambda, which reads the
content from S3.

This will allow us to handle content that is too big to fit in an SQS
message, and also make it easier to re-ingest content if we need to.

It also makes way for the sport.copy SES integration.